PreFinal

+++ Raspberry Pi/robot/CPPPort/StepperMotor.h
@@ -12,8 +12,8 @@ public:
     int portStep;
     int portDir;
 
-    int currentPosition = 0;
-    int neededPosition = 0;
+    int64_t currentPosition = 0;
+    int64_t neededPosition = 0;
     int stepDelay = 500;
 
     std::chrono::time_point<std::chrono::high_resolution_clock> lastTime;
